--- pages/login-loading.vue
@@ -1,13 +1,0 @@
-<script setup>
-
-
-
-</script>
-
-<template>
-  <div>
-    <h1>Loading</h1>
-  </div>
-</template>
-
-<style lang="scss" scoped></style>
